Refactor include paths and update project structure
Updated include paths in header files to reflect new directory
structure. Changed `VisitRecord` to inherit from `MedicalRecord`
instead of `IMedicalRecord`. Updated `domain.h` and `main.cpp`
to include headers from the new structure. Modified
`final_work_cpp.vcxproj` and `final_work_cpp.vcxproj.filters`
to include new paths and remove old ones.
