+++ data/MedicalRecord/visit_record.h
@@ -3,11 +3,11 @@
 
 #include "medical_record.h"
 
-class VisitRecord : public IMedicalRecord {
+class VisitRecord : public MedicalRecord {
 public:
-	VisitRecord(const string& d, const string& det) : IMedicalRecord(d, det) {}//VisitRecord based on MedicalRecord
+	VisitRecord(const string& d, const string& det) : MedicalRecord(d, det) {}//VisitRecord based on MedicalRecord
 
-	VisitRecord(const Date d, const string& det) : IMedicalRecord(d, det) {}
+	VisitRecord(const Date d, const string& det) : MedicalRecord(d, det) {}
 
 	void displayRecord() const override {
 		cout << "Visit Record - Date: " << _date.toString() << ", Details: " << _details << endl;
