+++ data/Users/user.h
@@ -1,8 +1,8 @@
 #ifndef USER_CLASS
 #define USER_CLASS
 
-#include "lib.h"
-#include "IUser.h"
+#include "../../domain/lib.h"
+#include "../Interfaces/IUser.h"
 
 class User : public IUser {
