--- domain.h
@@ -1,26 +1,0 @@
-//includes
-
-//lib
-#include "lib.h"
-
-
-//interfaces
-#include "IDataStorage.h"
-#include "IMedicalRecord.h"
-#include "IUser.h"
-
-//file storage
-#include "file_storage.h"
-
-//medical records
-#include "medical_record.h"
-#include "diagnostic_record.h"
-#include "treatment_record.h"
-#include "visit_record.h"
-
-//users
-#include "user.h"
-#include "admin.h"
-#include "doctor.h"
-#include "nurse.h"
-#include "admin_staff.h"
