+++ final_work_cpp.vcxproj
@@ -127,24 +127,24 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="main.cpp" />
+    <ClInclude Include="data\FileStorage\file_storage.h" />
+    <ClInclude Include="data\Interfaces\IDataStorage.h" />
+    <ClInclude Include="data\Interfaces\IMedicalRecord.h" />
+    <ClInclude Include="data\Interfaces\IUser.h" />
+    <ClInclude Include="data\MedicalRecord\diagnostic_record.h" />
+    <ClInclude Include="data\MedicalRecord\medical_record.h" />
+    <ClInclude Include="data\MedicalRecord\treatment_record.h" />
+    <ClInclude Include="data\MedicalRecord\visit_record.h" />
+    <ClInclude Include="data\Users\admin.h" />
+    <ClInclude Include="data\Users\admin_staff.h" />
+    <ClInclude Include="data\Users\doctor.h" />
+    <ClInclude Include="data\Users\nurse.h" />
+    <ClInclude Include="data\Users\user.h" />
+    <ClInclude Include="domain\domain.h" />
+    <ClInclude Include="domain\lib.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="admin.h" />
-    <ClInclude Include="admin_staff.h" />
-    <ClInclude Include="diagnostic_record.h" />
-    <ClInclude Include="doctor.h" />
-    <ClInclude Include="domain.h" />
-    <ClInclude Include="file_storage.h" />
-    <ClInclude Include="IDataStorage.h" />
-    <ClInclude Include="IMedicalRecord.h" />
-    <ClInclude Include="IUser.h" />
-    <ClInclude Include="lib.h" />
-    <ClInclude Include="medical_record.h" />
-    <ClInclude Include="nurse.h" />
-    <ClInclude Include="treatment_record.h" />
-    <ClInclude Include="user.h" />
-    <ClInclude Include="visit_record.h" />
+    <ClCompile Include="main.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
