+++ main.cpp
@@ -1,4 +1,4 @@
-#include "domain.h"
+#include "domain/domain.h"
 
 int main()
 {
